+++ usermgmt_gui/src/current_selected_view/configuration_state.rs
@@ -2,11 +2,11 @@ use std::path::PathBuf;
 
 use usermgmt_lib::config::LoadedMgmtConfig;
 
-use crate::{drawing::configuration::CacheForConfFiels, io_resource_manager::IoResourceManager};
+use crate::{drawing::configuration::CacheForConfFields, io_resource_manager::IoResourceManager};
 
 #[derive(Debug, Default)]
 pub struct ConfigurationState {
-    pub gui_field_cache: CacheForConfFiels,
+    pub gui_field_cache: CacheForConfFields,
     pub io_conf: IoResourceManager<LoadedMgmtConfig>,
     pub io_save_conf: IoResourceManager<PathBuf>,
 }