+++ usermgmt_gui/src/drawing/draw_utils.rs
@@ -1,3 +1,9 @@
+pub use group_drawing::GroupDrawing;
+pub use text_field_entry::TextFieldEntry;
+
+mod group_drawing;
+mod text_field_entry;
+
 use crate::prelude::*;
 
 use eframe::egui::{self, RichText};
@@ -8,84 +14,7 @@ use crate::{
     which_systems,
 };
 
-#[derive(Debug)]
-pub struct GroupDrawing<'a, 'b> {
-    name: &'a str,
-    tooltip: Option<&'b str>,
-}
-
-impl<'a, 'b> GroupDrawing<'a, 'b> {
-    pub fn new(name: &'a str) -> Self {
-        Self {
-            name,
-            tooltip: None,
-        }
-    }
-    pub fn with_tooltip(self, text: &'b str) -> GroupDrawing<'a, 'b> {
-        Self {
-            name: self.name,
-            tooltip: Some(text),
-        }
-    }
-    pub fn tooltip(self, text: Option<&'b str>) -> Self {
-        Self {
-            name: self.name,
-            tooltip: text,
-        }
-    }
-}
-
-#[derive(Debug)]
-enum ContentField<'a> {
-    Required(&'a mut String),
-    Optional(&'a mut Option<String>),
-}
-#[derive(Debug)]
-pub struct TextFieldEntry<'a, 'b> {
-    label: &'a str,
-    content: ContentField<'a>,
-    tool_tip: Option<&'b str>,
-    as_password: bool,
-}
-
-impl<'a, 'b> TextFieldEntry<'a, 'b> {
-    pub fn new(label: &'a str, content: &'a mut String) -> Self {
-        Self {
-            label,
-            content: ContentField::Required(content),
-            as_password: false,
-            tool_tip: None,
-        }
-    }
-    pub fn new_opt(label: &'a str, content: &'a mut Option<String>) -> Self {
-        Self {
-            label,
-            content: ContentField::Optional(content),
-            as_password: false,
-            tool_tip: None,
-        }
-    }
-    pub fn as_password(mut self) -> Self {
-        self.as_password = true;
-        self
-    }
-    pub fn tool_tip(self, too_tip: Option<&'b str>) -> Self {
-        Self {
-            label: self.label,
-            content: self.content,
-            as_password: self.as_password,
-            tool_tip: too_tip,
-        }
-    }
-    pub fn with_tooltip(self, tooltip: &'b str) -> Self {
-        Self {
-            label: self.label,
-            content: self.content,
-            as_password: self.as_password,
-            tool_tip: Some(tooltip),
-        }
-    }
-}
+use super::ProduceIoStatusMessages;
 
 pub fn tooltip_widget(ui: &mut egui::Ui, settings: &Settings, text: &str) {
     ui.label(
@@ -169,7 +98,7 @@ pub fn draw_ssh_credentials(
         ui,
         settings,
         &GroupDrawing::new(settings.texts().ssh_cred())
-            .with_tooltip(settings.tooltiptexts().ssh_creds()),
+            .add_tooltip(settings.tooltiptexts().ssh_creds()),
         username,
         password,
     );
@@ -186,7 +115,7 @@ pub fn draw_ldap_credentials(
         ui,
         settings,
         &GroupDrawing::new(settings.texts().ldap_cred())
-            .with_tooltip(settings.tooltiptexts().ldap_creds()),
+            .add_tooltip(settings.tooltiptexts().ldap_creds()),
         username,
         password,
     );
@@ -209,7 +138,7 @@ pub fn user_password_box(
             ui,
             settings,
             &mut TextFieldEntry::new_opt(settings.texts().password(), password_content)
-                .as_password(),
+                .with_as_password(),
         );
     });
 }
@@ -220,13 +149,13 @@ pub fn draw_box_group<R>(
     group: &GroupDrawing,
     on_draw: impl FnOnce(&mut egui::Ui) -> R,
 ) {
-    if let Some(tool_tip_name) = group.tooltip {
+    if let Some(tool_tip_name) = group.tooltip() {
         ui.horizontal(|ui| {
-            ui.label(RichText::new(group.name).strong());
+            ui.label(RichText::new(group.name()).strong());
             tooltip_widget(ui, settings, tool_tip_name);
         });
     } else {
-        ui.label(RichText::new(group.name).strong());
+        ui.label(RichText::new(group.name()).strong());
     }
     ui.group(on_draw);
 }
@@ -264,7 +193,7 @@ pub fn link_box(
     })
 }
 pub fn entry_field(ui: &mut egui::Ui, settings: &Settings, entry_field: &mut TextFieldEntry) {
-    draw_enty_field(ui, settings, entry_field)
+    draw_entry_field(ui, settings, entry_field)
 }
 pub fn whole_pos_number_fields<T>(
     ui: &mut egui::Ui,
@@ -276,7 +205,7 @@ pub fn whole_pos_number_fields<T>(
     T: ToPrimitive + FromPrimitive + Bounded + Copy,
 {
     let mut float: f32 = content.to_f32().unwrap_or_else(|| {
-        warn!("Interger value could not be casted to f32 for gui.");
+        warn!("Integer value could not be casted to f32 for gui.");
         warn!("Using the biggest possible f32 value instead.");
         f32::MAX
     });
@@ -291,8 +220,8 @@ pub fn whole_pos_number_fields<T>(
         let new_value = <T as FromPrimitive>::from_f32(rounded).unwrap_or_else(|| {
             let max = <T as Bounded>::max_value();
 
-            warn!("Floating value from gui could not be casted to interger value.");
-            warn!("Using biggest interger number as new value instead.");
+            warn!("Floating value from gui could not be casted to integer value.");
+            warn!("Using biggest integer number as new value instead.");
             max
         });
         *content = new_value;
@@ -309,7 +238,7 @@ pub fn whole_neg_number_fields<T>(
     T: ToPrimitive + FromPrimitive + Signed + Bounded + Copy,
 {
     let mut float: f32 = content.to_f32().unwrap_or_else(|| {
-        warn!("Interger value could not be casted to f32 for gui.");
+        warn!("Integer value could not be casted to f32 for gui.");
         if content.is_negative() {
             warn!("Using the smallest possible f32 value instead.");
             f32::MIN
@@ -327,12 +256,12 @@ pub fn whole_neg_number_fields<T>(
         let rounded = float.round();
         let new_value = <T as FromPrimitive>::from_f32(rounded).unwrap_or_else(|| {
             let (min, max) = (<T as Bounded>::min_value(), <T as Bounded>::max_value());
-            warn!("Floating value from gui could not be casted to interger value.");
+            warn!("Floating value from gui could not be casted to integer value.");
             if rounded.is_negative() {
-                warn!("Using smallest interger number as new value instead.");
+                warn!("Using smallest integer number as new value instead.");
                 min
             } else {
-                warn!("Using biggest interger number as new value instead.");
+                warn!("Using biggest integer number as new value instead.");
                 max
             }
         });
@@ -340,51 +269,30 @@ pub fn whole_neg_number_fields<T>(
     });
 }
 
-#[allow(clippy::too_many_arguments)]
-pub fn draw_status_msg_w_label<T>(
+pub fn draw_status_msg_w_label<T, C>(
     ui: &mut egui::Ui,
     settings: &Settings,
     label: &str,
     status: &IoTaskStatus<T>,
-    msg_init: impl FnOnce() -> String,
-    msg_loading: impl FnOnce() -> String,
-    msg_success: impl FnOnce(&T) -> String,
-    error_msg: impl FnOnce() -> String,
-) {
-    status_msg(
-        ui,
-        settings,
-        label,
-        status,
-        msg_init,
-        msg_loading,
-        msg_success,
-        error_msg,
-    )
+    msg: C,
+) where
+    C: ProduceIoStatusMessages<T>,
+{
+    status_msg(ui, settings, label, status, msg)
 }
 
-pub fn draw_status_msg<T>(
+pub fn draw_status_msg<T, C>(
     ui: &mut egui::Ui,
     settings: &Settings,
     status: &IoTaskStatus<T>,
-    msg_init: impl FnOnce() -> String,
-    msg_loading: impl FnOnce() -> String,
-    msg_success: impl FnOnce(&T) -> String,
-    error_msg: impl FnOnce() -> String,
-) {
-    status_msg(
-        ui,
-        settings,
-        settings.texts().general_status(),
-        status,
-        msg_init,
-        msg_loading,
-        msg_success,
-        error_msg,
-    )
+    msg: C,
+) where
+    C: ProduceIoStatusMessages<T>,
+{
+    status_msg(ui, settings, settings.texts().general_status(), status, msg)
 }
 
-pub fn draw_credentails(ui: &mut egui::Ui, window: &mut UsermgmtWindow, supports_dir: bool) {
+pub fn draw_credentials(ui: &mut egui::Ui, window: &mut UsermgmtWindow, supports_dir: bool) {
     which_systems::draw_which_system(ui, &window.settings, &mut window.which_sys, supports_dir);
     if window.is_ssh_cred_needed(supports_dir) {
         draw_ssh_credentials(ui, &window.settings, &mut window.ssh_state);
@@ -394,58 +302,43 @@ pub fn draw_credentails(ui: &mut egui::Ui, window: &mut UsermgmtWindow, supports
     }
 }
 
-fn draw_enty_field(ui: &mut egui::Ui, settings: &Settings, entry_field: &mut TextFieldEntry) {
+fn draw_entry_field(ui: &mut egui::Ui, settings: &Settings, entry_field: &mut TextFieldEntry) {
     ui.horizontal(|ui| {
-        ui.label(entry_field.label);
+        ui.label(entry_field.label());
 
-        let mut empty = String::default();
-        let mut opt = false;
-        let content: &mut String = match &mut entry_field.content {
-            ContentField::Required(content) => content,
-            ContentField::Optional(optional) => {
-                opt = true;
-                optional.as_mut().unwrap_or(&mut empty)
-            }
-        };
+        let password = entry_field.as_password();
+        let content = entry_field.content();
+        let mut buffer = content.to_owned();
         if ui
-            .add(egui::TextEdit::singleline(content).password(entry_field.as_password))
+            .add(egui::TextEdit::singleline(&mut buffer).password(password))
             .changed()
-            && opt
-            && !content.as_str().trim().is_empty()
         {
-            let content = content.to_owned();
-            if let ContentField::Optional(to_change) = &mut entry_field.content {
-                **to_change = Some(content)
-            } else {
-                unreachable!();
-            }
+            entry_field.set_content(buffer);
         }
-        if let Some(tool_tip_text) = entry_field.tool_tip {
+        if let Some(tool_tip_text) = entry_field.tool_tip() {
             tooltip_widget(ui, settings, tool_tip_text)
         }
     });
 }
 
-#[allow(clippy::too_many_arguments)]
-fn status_msg<T>(
+fn status_msg<T, C>(
     ui: &mut egui::Ui,
     settings: &Settings,
     label: &str,
     status: &IoTaskStatus<T>,
-    msg_init: impl FnOnce() -> String,
-    msg_loading: impl FnOnce() -> String,
-    msg_success: impl FnOnce(&T) -> String,
-    error_msg: impl FnOnce() -> String,
-) {
+    mut msg: C,
+) where
+    C: ProduceIoStatusMessages<T>,
+{
     draw_box_group(ui, settings, &GroupDrawing::new(label), |ui| {
         let colors = settings.colors();
         let (color, raw_text) = match status {
-            IoTaskStatus::NotStarted => (colors.init_msg(), msg_init()),
-            IoTaskStatus::Loading => (colors.loading_msg(), msg_loading()),
-            IoTaskStatus::Successful(val) => (colors.success_msg(), msg_success(val)),
+            IoTaskStatus::NotStarted => (colors.init_msg(), msg.msg_init()),
+            IoTaskStatus::Loading => (colors.loading_msg(), msg.msg_loading()),
+            IoTaskStatus::Successful(val) => (colors.success_msg(), msg.msg_success(val)),
             IoTaskStatus::Failed(error) => (
                 colors.err_msg(),
-                general_utils::error_status(&error_msg(), error),
+                general_utils::error_status(&msg.msg_error(), error),
             ),
         };
         let text = RichText::new(raw_text).color(color).strong();