+++ usermgmt_gui/src/drawing/modify_state.rs
@@ -1,13 +1,13 @@
 use usermgmt_lib::operations;
 
 use crate::{
-    current_selected_view::ModifyState, general_utils::PreparationBeforIoTask, prelude::*,
+    current_selected_view::ModifyState, general_utils::PreparationBeforeIoTask, prelude::*,
 };
 
 use super::draw_utils::{GroupDrawing, TextFieldEntry};
 pub fn draw(ui: &mut egui::Ui, window: &mut UsermgmtWindow) {
     draw_typing_fields(ui, &window.settings, &mut window.modify_state);
-    draw_utils::draw_credentails(ui, window, false);
+    draw_utils::draw_credentials(ui, window, false);
     ui.separator();
     if ui.button("Modify User").clicked() {
         handle_modify_req(window);
@@ -19,16 +19,18 @@ pub fn draw(ui: &mut egui::Ui, window: &mut UsermgmtWindow) {
         ui,
         &window.settings,
         window.modify_state.res_io.status(),
-        || text.modify_init().to_string(),
-        || format!("{} {}", text.modify_loading(), &last_username),
-        |username| format!("{} {}", text.modify_success(), username),
-        || format!("{} {}", text.modify_failure(), &last_username),
+        (
+            || text.modify_init().to_string(),
+            || format!("{} {}", text.modify_loading(), &last_username),
+            |username: &String| format!("{} {}", text.modify_success(), username),
+            || format!("{} {}", text.modify_failure(), &last_username),
+        ),
     );
 }
 
 fn handle_modify_req(window: &mut UsermgmtWindow) {
     window.modify_state.last_added_username = window.modify_state.username.clone();
-    if let Ok(PreparationBeforIoTask {
+    if let Ok(PreparationBeforeIoTask {
         ldap_cred,
         ssh_cred,
         config,
@@ -49,7 +51,7 @@ fn handle_modify_req(window: &mut UsermgmtWindow) {
                         )?;
                         Ok(username)
                     },
-                    String::from("Modifing User"),
+                    String::from("Modifying User"),
                 );
             }
             Err(error) => window.modify_state.res_io.set_error(error),
@@ -104,7 +106,7 @@ fn draw_typing_fields(ui: &mut egui::Ui, settings: &Settings, modify_state: &mut
             ui,
             settings,
             &mut modify_state.qos,
-            &GroupDrawing::new(texts.qos()).with_tooltip(tooltips.qos()),
+            &GroupDrawing::new(texts.qos()).add_tooltip(tooltips.qos()),
         );
     });
 }