+++ usermgmt_lib/src/app_error.rs
@@ -2,6 +2,8 @@ use crate::prelude::*;
 use anyhow::anyhow;
 use std::process::Output;
 
+/// Converts parameter `output` from spawned OS process into a result
+/// for easier error propagation as rust result.
 pub fn output_to_result(output: Output) -> AppResult<Output> {
     let status = output.status;
     if status.success() {